requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==2.3.2               # via kombu
ansible-runner==1.3.1
appdirs==1.4.2
argparse==1.4.0           # via uwsgitop
asgi-amqp==1.1.3
asgiref==1.1.2
asn1crypto==0.24.0        # via cryptography
attrs==18.2.0             # via automat, service-identity, twisted
autobahn==19.2.1          # via daphne
automat==0.7.0            # via twisted
billiard==3.5.0.5         # via celery
boto==2.47.0
celery==4.2.1
cffi==1.12.1              # via cryptography
channels==1.1.8
constantly==15.1.0        # via twisted
cryptography==2.5         # via pyopenssl
daphne==1.3.0
defusedxml==0.5.0
django-auth-ldap==1.7.0
django-cors-headers==2.4.0
django-crum==0.7.2
django-extensions==2.0.0
django-jsonfield==1.0.1
django-oauth-toolkit==1.1.3
django-pglocks==1.0.2
django-polymorphic==2.0.2
django-radius==1.3.3
django-solo==1.1.3
django-split-settings==0.3.0
django-taggit==0.22.2
django==1.11.16
djangorestframework-yaml==1.0.3
djangorestframework==3.7.7
future==0.16.0            # via django-radius
hyperlink==18.0.0         # via twisted
idna==2.8                 # via hyperlink
incremental==17.5.0       # via twisted
inflect==2.1.0            # via jaraco.itertools
irc==16.2
isodate==0.6.0            # via python3-saml
jaraco.classes==2.0       # via jaraco.collections
jaraco.collections==2.0   # via irc, jaraco.text
jaraco.functools==2.0     # via irc, jaraco.text, tempora
jaraco.itertools==4.4.1   # via irc
jaraco.logging==2.0       # via irc
jaraco.stream==2.0        # via irc
jaraco.text==2.0          # via irc, jaraco.collections
jinja2==2.10
jsonpickle==1.1           # via asgi-amqp
jsonschema==2.6.0
kombu==4.2.1              # via asgi-amqp, celery
lxml==4.3.1               # via xmlsec
markdown==2.6.11
markupsafe==1.1.0         # via jinja2
more-itertools==6.0.0     # via irc, jaraco.functools, jaraco.itertools
msgpack-python==0.5.6     # via asgi-amqp
netaddr==0.7.19           # via pyrad
oauthlib==2.0.6           # via django-oauth-toolkit, requests-oauthlib, social-auth-core
ordereddict==1.1
pexpect==4.6.0
pkgconfig==1.4.0          # via xmlsec
prometheus_client==0.6.0
psutil==5.4.3
psycopg2==2.7.3.2
ptyprocess==0.6.0         # via pexpect
pyasn1-modules==0.2.4     # via python-ldap, service-identity
pyasn1==0.4.5             # via pyasn1-modules, python-ldap, service-identity
pycparser==2.19           # via cffi
pygerduty==0.37.0
pyhamcrest==1.9.0         # via twisted
pyjwt==1.7.1              # via social-auth-core, twilio
pyopenssl==19.0.0         # via service-identity
pyparsing==2.2.0
pyrad==2.1                # via django-radius
pysocks==1.6.8            # via twilio
python-daemon==2.2.0      # via ansible-runner
python-dateutil==2.7.2
python-ldap==3.1.0        # via django-auth-ldap
python-memcached==1.59
python-radius==1.0
python3-openid==3.1.0     # via social-auth-core
python3-saml==1.4.0
pytz==2018.9              # via celery, django, irc, tempora, twilio
pyyaml==3.13              # via djangorestframework-yaml
requests-futures==0.9.7
requests-oauthlib==1.2.0  # via social-auth-core
requests[security]==2.15.1
service-identity==17.0.0
simplejson==3.16.0        # via uwsgitop
six==1.12.0               # via asgi-amqp, asgiref, autobahn, automat, cryptography, django-extensions, irc, isodate, jaraco.classes, jaraco.collections, jaraco.itertools, jaraco.logging, jaraco.stream, pygerduty, pyhamcrest, pyopenssl, pyrad, python-dateutil, python-memcached, slackclient, social-auth-app-django, social-auth-core, tacacs-plus, tempora, twilio, txaio, websocket-client
slackclient==1.1.2
social-auth-app-django==2.1.0
social-auth-core==3.0.0
tacacs_plus==1.0
tempora==1.14             # via irc, jaraco.logging
twilio==6.10.4
twisted==18.9.0           # via daphne
txaio==18.8.1             # via autobahn
typing==3.6.6             # via django-extensions
uwsgi==2.0.17
uwsgitop==0.10.0
vine==1.2.0               # via amqp
websocket-client==0.47.0  # via slackclient
xmlsec==1.3.3             # via python3-saml
zope.interface==4.6.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
